SCH-2933 Refactor payment method into a couple of components
