+++ components/src/components/elements/payment-method/index.ts
@@ -1,1 +1,2 @@
 export * from "./PaymentMethod";
+export * from "./PaymentMethodDetails";
