+++ components/src/components/shared/index.ts
@@ -1,4 +1,5 @@
 export * from "./checkout-dialog";
+export * from "./payment-dialog";
 export * from "./payment-form";
 export * from "./period-toggle";
 export * from "./unsubscribe-dialog";
