+++ components/src/components/shared/payment-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./PaymentDialog";
